Change vim-poweline in favor of vim-airline
I added vim-airline as my default status bar. I also changed my default font to Droid Sans Mono and added pretty icons to the status bar. Instructions on how to set the font are in the README. I will *maybe* add installation instructions as part of the Makefile to ease this step.