Add FZF
Add fuzzy finder system wide using fzf installed from homebrew and the
Vim fzf plugin. Add some system wide configurations to make use of rg
(ripgrep) for grepping files.

Also add a function for fuzzy finding on open buffers. There's currently
an issue with this function as for some reason fzf_action is not working
with it; other than that everything is up to par with the CtrlP
functionality that I used.